Added file for arxiv plots using ToO notebook
The added notebook simulates the "GW170817 best fit polar for theta=26deg" kilonova model using the Golden Strategy for the 3 different cases as per the ToO report. This notebook was used to generate the Bright Kilonova plots used in the Arxiv/published format of the Rubin ToO 2024 workshop report. The python source file contains a new function defined to show the detections and non-detections in a data frame format for each strategy.